FIX: quvi-get: Use "default" with --print-streams
When there is only one media stream, and that is unnamed by the libquvi
media script, use "default" to identify this stream in the output.

We use "default" to identify the stream for the clarity of the users,
technically the string could be anything, e.g. "--stream foo" would
still select the default stream if there is only one stream available,
or "foo" matched none of the stream IDs.

Signed-off-by: Toni Gundogdu <legatvs@gmail.com>