FIX: Keep the input URL order
This fixes what commit 4d6aac87dde312ae0a8ee1a329426e01f40f746a was
originally intended to fix.  Without this, the input URL order gets
mangled which can be frustrating when processing URLs in batches and
operating over an unreliable connection.

Rather than reversing the order in the "check support" facility which is
called *foreach* input URL, reverse the lists only once *after* the
facility returns the results.

See the comment in the committed code for the reason we reverse the
order.

Signed-off-by: Toni Gundogdu <legatvs@gmail.com>