status: Replace use of GDateTime with static frames
Status update messages come now with fun new animation that everyone can
enjoy.  This replaces the use of GDateTime in the status update messages
which was introduced in quvi v0.9.2.

GDateTime was added in GLib 2.26, Debian Squeeze packages 2.24. This
makes it it possible to compile quvi on Squeeze, once again.

Other:
* Remove the "with_time" flag, check the string length instead

See also:
* http://article.gmane.org/gmane.comp.web.flash.quvi/293

Signed-off-by: Toni Gundogdu <legatvs@gmail.com>