Omit status updates unless --print-status is given
The program will now omit the status update messages from the stderr,
making it easier to parse the error messages from the output. This is
now the new default behaviour.

  - Implement "--print-status,-z"
    - Allow the users to turn on the status update messages
    - Allow configuring as core.print-status

Ideally, the program would print these messages by default, but given
their trivial nature, they might as well be off and turned easily on if
needed.  This makes it possible to keep the current verbosity system in
place without additional tinkering.

See also:
  https://www.mail-archive.com/quvi-devel@lists.sourceforge.net/msg00372.html

Signed-off-by: Toni Gundogdu <legatvs@gmail.com>